+++ src/sagemaker/xgboost/model.py
@@ -16,10 +16,11 @@ from __future__ import absolute_import
 import logging
 
 import sagemaker
+from sagemaker.deserializers import CSVDeserializer
 from sagemaker.fw_utils import model_code_key_prefix
 from sagemaker.fw_registry import default_framework_uri
 from sagemaker.model import FrameworkModel, MODEL_SERVER_WORKERS_PARAM_NAME
-from sagemaker.predictor import Predictor, npy_serializer, csv_deserializer
+from sagemaker.predictor import Predictor, npy_serializer
 from sagemaker.xgboost.defaults import XGBOOST_NAME
 
 logger = logging.getLogger("sagemaker")
@@ -42,7 +43,7 @@ class XGBoostPredictor(Predictor):
                 chain.
         """
         super(XGBoostPredictor, self).__init__(
-            endpoint_name, sagemaker_session, npy_serializer, csv_deserializer
+            endpoint_name, sagemaker_session, npy_serializer, CSVDeserializer()
         )
