+++ tests/unit/test_predictor.py
@@ -25,7 +25,6 @@ from sagemaker.predictor import (
     json_serializer,
     json_deserializer,
     csv_serializer,
-    csv_deserializer,
     StringDeserializer,
     StreamDeserializer,
     numpy_deserializer,
@@ -148,21 +147,6 @@ def test_csv_serializer_csv_reader():
         csv_file.seek(0)
         result = csv_serializer(csv_file)
         assert result == validation_data
-
-
-def test_csv_deserializer_single_element():
-    result = csv_deserializer(io.BytesIO(b"1"), "text/csv")
-    assert result == [["1"]]
-
-
-def test_csv_deserializer_array():
-    result = csv_deserializer(io.BytesIO(b"1,2,3"), "text/csv")
-    assert result == [["1", "2", "3"]]
-
-
-def test_csv_deserializer_2dimensional():
-    result = csv_deserializer(io.BytesIO(b"1,2,3\n3,4,5"), "text/csv")
-    assert result == [["1", "2", "3"], ["3", "4", "5"]]
 
 
 def test_json_deserializer_array():
